feat: add search for new configs after training ends
Previously it would get the list of trainings to do at the start, but
any modifications do the directory would not be reflected in memory. The
files are now fetched as needed rather than once at the start.

Failed jobs will also be skipped and remembered in to prevent infinite
looping.